Rename calendar components
